# Conflicts:
#	onePosition/app/src/main/java/cn/shopin/oneposition/fragments/moviefrag/collection/CollectionFrag.java
#	onePosition/app/src/main/res/layout/frag_balloon.xml
#	onePosition/app/src/main/res/layout/frag_collection.xml
#	onePosition/app/src/main/res/layout/frag_movie.xml

+++ onePosition/app/src/main/java/cn/shopin/oneposition/entity/movie/BannerDetailEntity.java
@@ -42,7 +42,7 @@ public class BannerDetailEntity {
     private int category;
     private int commentcount;
     private int sharecount;
-    private int prize;
+    private float prize;
     private int wantwatchnums;
     private int unwantwatchnums;
     private int picsCount;
@@ -178,11 +178,11 @@ public class BannerDetailEntity {
         this.sharecount = sharecount;
     }
 
-    public int getPrize() {
+    public float getPrize() {
         return prize;
     }
 
-    public void setPrize(int prize) {
+    public void setPrize(float prize) {
         this.prize = prize;
     }
